Add function to find SDF to foil
The foil is described by a cubic Bezier, but this makes the signed distance function tricky to compute.

For any 2D point X in space, you need to find the closest 1D point t along the curve. Since the tangent at the correct t will be perpendicular to X-curve(t), their dot-product is zero and this can be posed as a root-finding problem. However, it's a quintic polynomial in t, which are hard=>impossible to solve analytically!

To avoid this, I used Roots.find_zeros() which tries to find all the roots in a function between a range. I add the endpoints t=0,1 to whatever this function finds and call these the candidate t values. Then I find the candidate with the smallest absolute distance.

Julia makes this only 7 lines of code, but it's hiding quite a bit of work in find_roots(), so it's pretty slow. Still, even using a coarse tolerance the results look very good!